+++ services/workflows-service/scripts/workflows/ui-definition/kyb-with-associated-companies/definition/compose-kyb-with-associated-companies-definition.ts
@@ -188,7 +188,6 @@ export const composeKybWithAssociatedCompaniesDefinition = ({
     },
     config: {
       isExample: true,
-      createCollectionFlowToken: true,
       workflowLevelResolution: true,
       childCallbackResults: [
         {