+++ services/workflows-service/scripts/workflows/ui-definition/kyb-with-associated-companies/definition/compose-kyc-child-workflow-definition.ts
@@ -57,7 +57,6 @@ export const composeKycChildWorkflowDefinition = ({
       apiPlugins: [],
     },
     config: {
-      createCollectionFlowToken: true,
       workflowLevelResolution: true,
       isCaseOverviewEnabled: true,
     },