+++ examples/headless-example/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @ballerine/headless-example
 
+## 0.3.70
+
 ## 0.3.68
 
 ### Patch Changes
@@ -7,6 +9,19 @@
 - Updated dependencies
   - @ballerine/common@0.9.54
   - @ballerine/workflow-browser-sdk@0.6.69
+  - @ballerine/workflow-browser-sdk@0.6.71
+
+## 0.3.69
+
+### Patch Changes
+
+- @ballerine/workflow-browser-sdk@0.6.70
+
+## 0.3.68
+
+### Patch Changes
+
+- @ballerine/workflow-browser-sdk@0.6.69
 
 ## 0.3.67