+++ sdks/workflow-browser-sdk/CHANGELOG.md
@@ -1,11 +1,27 @@
 # @ballerine/workflow-browser-sdk
 
+## 0.6.71
+
 ## 0.6.69
 
 ### Patch Changes
 
 - Updated dependencies
   - @ballerine/common@0.9.54
+  - @ballerine/workflow-core@0.6.71
+
+## 0.6.70
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.70
+
+## 0.6.69
+
+### Patch Changes
+
+- Updated dependencies
   - @ballerine/workflow-core@0.6.69
 
 ## 0.6.68
@@ -36,6 +52,7 @@
 
 ### Patch Changes
 
+- Added additionalContext
 - version bump
 - Updated dependencies
   - @ballerine/workflow-core@0.6.65